- auth based on database

+++ app/set.go
@@ -22,7 +22,7 @@ type set struct {
 func (s *Server) handleGetSet() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// auth
-		_, err := validateToken(w, r)
+		claims, err := validateToken(w, r)
 		if err != nil {
 			return
 		}
@@ -37,26 +37,26 @@ func (s *Server) handleGetSet() http.HandlerFunc {
 		}
 
 		set := set{ID: id}
-		if err := set.getSet(s.DB); err != nil {
+		if err := set.getSet(s.DB, claims.UserID); err != nil {
 			switch err {
 			case sql.ErrNoRows:
 				log.Println(err.Error())
 				respondWithError(w, http.StatusNotFound, "Set not found")
 			default:
 				log.Println(err.Error())
-				respondWithError(w, http.StatusInternalServerError, err.Error())
+				respondWithError(w, http.StatusInternalServerError, "Internal server error")
 			}
 			return
 		}
 
-		respondWithJSON(w, http.StatusOK, s)
+		respondWithJSON(w, http.StatusOK, set)
 	}
 }
 
 func (s *Server) handleGetSets() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// auth
-		_, err := validateToken(w, r)
+		claims, err := validateToken(w, r)
 		if err != nil {
 			return
 		}
@@ -73,10 +73,10 @@ func (s *Server) handleGetSets() http.HandlerFunc {
 			start = 0
 		}
 
-		sets, err := set.getSets(s.DB, start, count)
+		sets, err := set.getSets(s.DB, start, count, claims.UserID)
 		if err != nil {
 			log.Println(err.Error())
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			respondWithError(w, http.StatusInternalServerError, "Internal server error")
 			return
 		}
 
@@ -87,7 +87,7 @@ func (s *Server) handleGetSets() http.HandlerFunc {
 func (s *Server) handleCreateSet() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// auth
-		_, err := validateToken(w, r)
+		claims, err := validateToken(w, r)
 		if err != nil {
 			return
 		}
@@ -102,9 +102,9 @@ func (s *Server) handleCreateSet() http.HandlerFunc {
 		}
 		defer r.Body.Close()
 
-		if err := set.createSet(s.DB); err != nil {
+		if err := set.createSet(s.DB, claims.UserID); err != nil {
 			log.Println(err.Error())
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			respondWithError(w, http.StatusInternalServerError, "Internal server error")
 			return
 		}
 
@@ -115,7 +115,7 @@ func (s *Server) handleCreateSet() http.HandlerFunc {
 func (s *Server) handleUpdateSet() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// auth
-		_, err := validateToken(w, r)
+		claims, err := validateToken(w, r)
 		if err != nil {
 			return
 		}
@@ -139,9 +139,9 @@ func (s *Server) handleUpdateSet() http.HandlerFunc {
 		defer r.Body.Close()
 		set.ID = id
 
-		if err := set.updateSet(s.DB); err != nil {
+		if err := set.updateSet(s.DB, claims.UserID); err != nil {
 			log.Println(err.Error())
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			respondWithError(w, http.StatusInternalServerError, "Internal server error")
 			return
 		}
 
@@ -153,7 +153,7 @@ func (s *Server) handleUpdateSet() http.HandlerFunc {
 func (s *Server) handleDeleteSet() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// auth
-		_, err := validateToken(w, r)
+		claims, err := validateToken(w, r)
 		if err != nil {
 			return
 		}
@@ -168,9 +168,9 @@ func (s *Server) handleDeleteSet() http.HandlerFunc {
 		}
 
 		set := set{ID: id}
-		if err := set.deleteSet(s.DB); err != nil {
+		if err := set.deleteSet(s.DB, claims.UserID); err != nil {
 			log.Println(err.Error())
-			respondWithError(w, http.StatusInternalServerError, err.Error())
+			respondWithError(w, http.StatusInternalServerError, "Internal server error")
 			return
 		}
 
@@ -179,16 +179,16 @@ func (s *Server) handleDeleteSet() http.HandlerFunc {
 }
 
 // getSet fetches a set from database with id
-func (s *set) getSet(db *sql.DB) error {
-	return db.QueryRow("SELECT user_id, weight, exercise, repetitions FROM sets WHERE id=$1",
-		s.ID).Scan(&s.UserID, &s.Weight, &s.Exercise, &s.Repetitions)
+func (s *set) getSet(db *sql.DB, userID string) error {
+	return db.QueryRow("SELECT user_id, weight, exercise, repetitions FROM sets WHERE id=$1 AND user_id=$2",
+		s.ID, userID).Scan(&s.UserID, &s.Weight, &s.Exercise, &s.Repetitions)
 }
 
 // getSets fetches multiple sets from database with user id
-func (s *set) getSets(db *sql.DB, start, count int) ([]set, error) {
+func (s *set) getSets(db *sql.DB, start, count int, userID string) ([]set, error) {
 	rows, err := db.Query(
-		"SELECT id, user_id, weight, exercise, repetitions FROM sets LIMIT $1 OFFSET $2",
-		count, start)
+		"SELECT id, user_id, weight, exercise, repetitions FROM sets WHERE user_id=$1 LIMIT $2 OFFSET $3",
+		userID, count, start)
 
 	if err != nil {
 		return nil, err
@@ -210,26 +210,26 @@ func (s *set) getSets(db *sql.DB, start, count int) ([]set, error) {
 }
 
 // updateSet executes update query to database
-func (s *set) updateSet(db *sql.DB) error {
+func (s *set) updateSet(db *sql.DB, userID string) error {
 	_, err :=
-		db.Exec("UPDATE sets SET user_id=$2, weight=$3, exercise=$4, repetitions=$5 WHERE id=$1",
-			s.ID, s.UserID, s.Weight, s.Exercise, s.Repetitions)
+		db.Exec("UPDATE sets SET user_id=$2, weight=$3, exercise=$4, repetitions=$5 WHERE id=$1 AND user_id=$2",
+			s.ID, userID, s.Weight, s.Exercise, s.Repetitions)
 
 	return err
 }
 
 // deleteSet deletes a set from database with
-func (s *set) deleteSet(db *sql.DB) error {
-	_, err := db.Exec("DELETE FROM sets WHERE id=$1", s.ID)
+func (s *set) deleteSet(db *sql.DB, userID string) error {
+	_, err := db.Exec("DELETE FROM sets WHERE id=$1 and user_id=$2", s.ID, userID)
 
 	return err
 }
 
 // createSet creates a set into database with given JSON
-func (s *set) createSet(db *sql.DB) error {
+func (s *set) createSet(db *sql.DB, userID string) error {
 	err := db.QueryRow(
 		"INSERT INTO sets(user_id, weight, exercise, repetitions) VALUES($1, $2, $3, $4) RETURNING id",
-		s.UserID, s.Weight, s.Exercise, s.Repetitions).Scan(&s.ID)
+		userID, s.Weight, s.Exercise, s.Repetitions).Scan(&s.ID)
 
 	if err != nil {
 		return err
